feat(ui): inline Settings panel as third sidebar surface (#479 slice 2)
Slice 2 of the Settings-window-deletion. Settings is now an
inline panel rendered inside the main app shell, not a separate
window. Click the Settings sidebar item, palette's
\"Open Settings: …\" actions, the FirstRunModal callbacks, and
the dictation section's deep-link to Permissions all flip
\`activeSection = \"settings\"\` (+ the requested tab) instead of
opening the standalone window.

## What's new

- \`SettingsPanel.svelte\` — extracted from the
  \`/settings\` route. Owns the tab strip, model picker state +
  lifecycle, and the cross-context \`settings:goto-tab\` listener.
  Bindable \`activeTab\` prop so the parent can deep-link in.
- \`/settings\` route reduced to a thin wrapper that mounts
  \`<SettingsPanel/>\`. Slice 3 will delete the route entirely
  along with the Tauri window definition + capability + IPC.
- \`+page.svelte\`:
  - \`activeSection = \"settings\"\` mounts \`<SettingsPanel
    bind:activeTab={settingsActiveTab} />\` inside the app
    shell.
  - \`onSidebarSelect\` for \`settings\` flips \`activeSection\` (no
    longer opens a window).
  - \`openSettingsTab(tab)\` is now a synchronous local helper
    that flips \`settingsActiveTab\` + \`activeSection\`. All
    seven palette actions, the model-picker deep-link, and the
    Permissions deep-link route through it unchanged at the
    callsite.

## Tests

- \`command-palette.spec.ts\` 'clicking a Settings row dispatches
  open_settings' rewritten as 'swaps to the Settings panel' —
  asserts the requested tab is \`aria-current=\"page\"\` after the
  palette row click. The pre-r2 IPC-call assertion is gone
  because the IPC isn't fired anymore.
- 84 specs pass.

## Slice 1 → 2 → 3 → 4

- 1 (#480 merged) — sidebar shell, panel switching
- 2 (this) — Settings inlined as third panel
- 3 — delete \`/settings\` route + \`settings_window\` Rust module
  + \`capabilities/settings.json\` + \`tauri.conf.json\` window +
  \`open_settings\` IPC + the tray menu's window-show path
- 4 — vibe pass: tooltips, hover, motion, possibly real icons
